practice_003d_kafka_connect_cdc/app/pyproject.toml
[project]
name = "kafka-connect-cdc"
version = "0.1.0"
description = "Practice 003d: Kafka Connect & Change Data Capture"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.6.0",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.38.0",
]
